Updated angular-facebook.js
Now, function passed to Facebook service methods through arguments is
not expected to be the last argument neccesarily, instead, it is
replaced in the same position it was passed.

This is needed in some cases like for the login method, which
accordingly to the Facebook API documentation, you can optionally pass
an object as last argument with extra options.